extract connectivityRequestHandler
